% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{make_wvar_mimu_obj}
\alias{make_wvar_mimu_obj}
\title{TO DO}
\usage{
make_wvar_mimu_obj(..., freq, unit, sensor.name, exp.name)
}
\description{
TO DO
}
\examples{
n = 10^6
Xt = rnorm(n/4)
Yt = rnorm(n/2) + cumsum(rnorm(n/2, 0, 10^(-2)))
Zt = rnorm(n) + cumsum(rnorm(n, 0, 10^(-3)))
obj = make_wvar_mimu_obj(Xt, Yt, Zt, freq = 100, unit = "s",
sensor.name = "MTiG - Gyro. X", exp.name = c("today", "yesterday", "a few days ago"))
plot(obj)
}
